+++ packages/@expo/cli/bin/cli.ts
@@ -25,7 +25,7 @@ const commands: { [command: string]: () => Promise<Command> } = {
   config: () => import('../src/config').then((i) => i.expoConfig),
   export: () => import('../src/export').then((i) => i.expoExport),
   'export:web': () => import('../src/export/web').then((i) => i.expoExportWeb),
-  'bundle:proxy': () => import('../src/bundle-proxy').then((i) => i.expoBundleProxy),
+  'export:embed': () => import('../src/export/embed').then((i) => i.expoExportEmbed),
 
   // Auxiliary commands
   install: () => import('../src/install').then((i) => i.expoInstall),
@@ -85,7 +85,7 @@ if (!isSubcommand && args['--help']) {
     'run:ios': runIos,
     'run:android': runAndroid,
     // Don't document this command
-    'bundle:proxy': bundleProxy_unused,
+    'export:embed': bundleProxy_unused,
     ...others
   } = commands;