+++ packages/@expo/cli/src/export/embed/exportEmbedAsync.ts
@@ -1,8 +1,8 @@
-import { loadMetroConfigAsync } from '../start/server/metro/instantiateMetro';
-import { importCliBuildBundleWithConfigFromProject } from '../start/server/metro/resolveFromProject';
+import { loadMetroConfigAsync } from '../../start/server/metro/instantiateMetro';
+import { importCliBuildBundleWithConfigFromProject } from '../../start/server/metro/resolveFromProject';
 import { Options } from './resolveOptions';
 
-export async function bundleProxyAsync(projectRoot: string, options: Options) {
+export async function exportEmbedAsync(projectRoot: string, options: Options) {
   const { config } = await loadMetroConfigAsync(projectRoot, {
     maxWorkers: options.maxWorkers,
     resetCache: options.resetCache,